refactor: mypy

+++ libs/ktem/ktem/llms/manager.py
@@ -1,4 +1,4 @@
-from typing import Optional, Type, overload
+from typing import Any, overload
 
 from sqlalchemy import select
 from sqlalchemy.orm import Session
@@ -13,11 +13,11 @@ from .db import LLMTable, engine
 class LLMManager:
     """Represent a pool of models"""
 
-    def __init__(self):
+    def __init__(self) -> None:
         self._models: dict[str, ChatLLM] = {}
-        self._info: dict[str, dict] = {}
+        self._info: dict[str, dict[str, Any]] = {}
         self._default: str = ""
-        self._vendors: list[Type] = []
+        self._vendors: list[type[ChatLLM]] = []
 
         if hasattr(flowsettings, "KH_LLMS"):
             for name, model in flowsettings.KH_LLMS.items():
@@ -36,7 +36,7 @@ class LLMManager:
         self.load()
         self.load_vendors()
 
-    def load(self):
+    def load(self) -> None:
         """Load the model pool from database"""
         self._models, self._info, self._default = {}, {}, ""
         with Session(engine) as session:
@@ -53,7 +53,7 @@ class LLMManager:
                 if item.default:
                     self._default = item.name
 
-    def load_vendors(self):
+    def load_vendors(self) -> None:
         from kotaemon.llms import (
             AzureChatOpenAI,
             ChatOpenAI,
@@ -86,18 +86,16 @@ class LLMManager:
         return key in self._models
 
     @overload
-    def get(self, key: str, default: None) -> Optional[ChatLLM]:
-        ...
+    def get(self, key: str, default: None) -> ChatLLM | None: ...
 
     @overload
-    def get(self, key: str, default: ChatLLM) -> ChatLLM:
-        ...
+    def get(self, key: str, default: ChatLLM) -> ChatLLM: ...
 
-    def get(self, key: str, default: Optional[ChatLLM] = None) -> Optional[ChatLLM]:
+    def get(self, key: str, default: ChatLLM | None = None) -> ChatLLM | None:
         """Get model by name with default value"""
         return self._models.get(key, default)
 
-    def settings(self) -> dict:
+    def settings(self) -> dict[str, Any]:
         """Present model pools option for gradio"""
         return {
             "label": "LLM",
@@ -105,7 +103,7 @@ class LLMManager:
             "value": self.get_default_name(),
         }
 
-    def options(self) -> dict:
+    def options(self) -> dict[str, ChatLLM]:
         """Present a dict of models"""
         return self._models
 
@@ -115,10 +113,10 @@ class LLMManager:
         Returns:
             str: random model name in the pool
         """
-        import random
-
         if not self._models:
             raise ValueError("No models in pool")
+
+        import random
 
         return random.choice(list(self._models.keys()))
 
@@ -154,11 +152,11 @@ class LLMManager:
         """
         return self._models[self.get_default_name()]
 
-    def info(self) -> dict:
+    def info(self) -> dict[str, dict[str, Any]]:
         """List all models"""
         return self._info
 
-    def add(self, name: str, spec: dict, default: bool):
+    def add(self, name: str, spec: dict[str, Any], *, default: bool) -> None:
         """Add a new model to the pool"""
         name = name.strip()
         if not name:
@@ -166,7 +164,6 @@ class LLMManager:
 
         try:
             with Session(engine) as session:
-
                 if default:
                     # turn all models to non-default
                     session.query(LLMTable).update({"default": False})
@@ -176,30 +173,34 @@ class LLMManager:
                 session.add(item)
                 session.commit()
         except Exception as e:
-            raise ValueError(f"Failed to add model {name}: {e}")
+            raise ValueError(f"Failed to add model {name}: {e}") from e
 
         self.load()
 
-    def delete(self, name: str):
+    def delete(self, name: str) -> None:
         """Delete a model from the pool"""
+        if not name:
+            raise ValueError("Name must not be empty")
+
         try:
             with Session(engine) as session:
                 item = session.query(LLMTable).filter_by(name=name).first()
+                if not item:
+                    raise ValueError(f"Model {name} not found")
                 session.delete(item)
                 session.commit()
         except Exception as e:
-            raise ValueError(f"Failed to delete model {name}: {e}")
+            raise ValueError(f"Failed to delete model {name}: {e}") from e
 
         self.load()
 
-    def update(self, name: str, spec: dict, default: bool):
+    def update(self, name: str, spec: dict[str, Any], *, default: bool) -> None:
         """Update a model in the pool"""
         if not name:
             raise ValueError("Name must not be empty")
 
         try:
             with Session(engine) as session:
-
                 if default:
                     # turn all models to non-default
                     session.query(LLMTable).update({"default": False})
@@ -212,11 +213,11 @@ class LLMManager:
                 item.default = default
                 session.commit()
         except Exception as e:
-            raise ValueError(f"Failed to update model {name}: {e}")
+            raise ValueError(f"Failed to update model {name}: {e}") from e
 
         self.load()
 
-    def vendors(self) -> dict:
+    def vendors(self) -> dict[str, type[ChatLLM]]:
         """Return list of vendors"""
         return {vendor.__qualname__: vendor for vendor in self._vendors}
